requirements.txt
alembic==1.18.4
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dask==2026.1.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
docker==7.1.0
dynaconf==3.2.12
entrypoints==0.4
evidently==0.4.22
executing==2.2.1
Faker==40.5.1
fastapi==0.110.0
fastjsonschema==2.21.2
feast==0.38.0
filelock==3.24.3
Flask==3.1.3
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
great-expectations==0.18.12
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
imbalanced-learn==0.12.0
importlib_metadata==7.2.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kafka-python==2.0.2
kiwisolver==1.4.9
lark==1.3.1
librt==0.8.1
lightgbm==4.3.0
litestar==2.21.0
litestar-htmx==0.5.0
locket==1.0.0
makefun==1.16.0
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mlflow==2.11.1
mmh3==5.2.0
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.0
mypy==1.19.1
mypy-protobuf==5.0.0
mypy_extensions==1.1.0
narwhals==2.17.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.3
notebook==7.5.4
notebook_shim==0.2.4
numpy==1.26.4
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.6
partd==1.4.2
pathspec==1.0.4
patsy==1.0.2
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
polyfactory==3.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
pure_eval==0.2.3
pyarrow==15.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
pytz==2024.2
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
querystring-parser==1.2.4
redis==5.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rich-click==1.9.7
rpds-py==0.30.0
ruamel.yaml==0.17.17
scikit-learn==1.4.1.post1
scipy==1.17.1
Send2Trash==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.47
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.6
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typeguard==4.5.1
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.29.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xgboost==2.0.3
zipp==3.23.0
